requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.5.7
charset-normalizer==3.2.0
idna==3.4
isort==5.12.0
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3
requests==2.31.0
six==1.16.0
tornado==6.3.2
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
